Remove azure_rke1_cluster module from Windows repository
The point of creating the `azure_rke1_cluster` module in this repository was so that the Windows team could look back on RKE1 and figure out how to debug legacy setups.

However, since RKE1 has been EOL for long enough and [the backing service for the MicrosoftDockerProvider has been shut down as of May 23rd, 2023](https://github.com/OneGet/MicrosoftDockerProvider), this module is now safe to remove from this repository.

It will live on in our Git history for future reference.